bot-tkinter-atividade-13/requirements.txt
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
botcity-maestro-sdk==0.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
comtypes==1.4.8
distro==1.9.0
idna==3.10
importlib_metadata==8.5.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
psutil==6.1.0
pynput==1.7.7
pyperclip==1.9.0
python-xlib==0.33
pywin32==308
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.3
zipp==3.20.2
